feat: add perception analyer 2d
