+++ perception_eval/perception_eval/tool/__init__.py
@@ -1,11 +1,18 @@
+from typing import Union
+
 from .gmm import Gmm
 from .gmm import load_sample
-from .perception_performance_analyzer import PerceptionPerformanceAnalyzer
+from .perception_analyzer2d import PerceptionAnalyzer2D
+from .perception_analyzer3d import PerceptionAnalyzer3D
 from .utils import MatchingStatus
 from .utils import PlotAxes
 
+# type aliases
+PerceptionAnalyzerType = Union[PerceptionAnalyzer2D, PerceptionAnalyzer3D]
+
 __all__ = (
-    "PerceptionPerformanceAnalyzer",
+    "PerceptionAnalyzer2D",
+    "PerceptionAnalyzer3D",
     "PlotAxes",
     "MatchingStatus",
     "Gmm",
